Registration with mail, migration with flyway,

+++ src/main/java/com/example/sweater/domain/Message.java
@@ -1,12 +1,14 @@
 package com.example.sweater.domain;
 
+import org.hibernate.annotations.GenericGenerator;
+
 import javax.persistence.*;
 
 @Entity// This tells Hibernate to make a table out of this class
 public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String text;
     private String tag;
@@ -30,11 +32,11 @@ public class Message {
         return author != null ? author.getUsername(): "<none>";
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
